ADS7830: Changed message name, refactor and tidy
