+++ src/drivers/adc/ads7830/ADS7830.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (C) 2020 PX4 Development Team. All rights reserved.
+ *   Copyright (C) 2025 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -32,7 +32,85 @@
  ****************************************************************************/
 
 #include "ADS7830.h"
+#include <px4_platform_common/module.h>
+#include <drivers/drv_adc.h>
 #include <cassert>
+
+ADS7830::ADS7830(const I2CSPIDriverConfig &config) :
+	I2C(config),
+	I2CSPIDriver(config),
+	_cycle_perf(perf_alloc(PC_ELAPSED, MODULE_NAME": cycle")),
+	_comms_errors(perf_alloc(PC_COUNT, MODULE_NAME": comms errors"))
+
+{
+	_regulator_report.device_id = get_device_id();
+	_regulator_report.resolution = 256; // 8-bit ADC
+	_regulator_report.v_ref = 5.0f; // Internal reference is 2.5V we are using 5V external ref via REF pin
+
+	for (unsigned i = 0; i < PX4_MAX_ADC_CHANNELS; ++i) {
+		_regulator_report.channel_id[i] = -1;
+	}
+}
+
+ADS7830::~ADS7830()
+{
+	ScheduleClear();
+	perf_free(_cycle_perf);
+	perf_free(_comms_errors);
+
+}
+
+void ADS7830::RunImpl()
+{
+	if (should_exit()) {
+		return;
+	}
+
+	perf_begin(_cycle_perf);
+	_regulator_report.timestamp = hrt_absolute_time();
+
+	int16_t value = 0;
+	ChannelSelection ch = cycleMeasure(&value);
+
+	if (ch != Invalid) {
+		_last_successful_measurement = hrt_absolute_time(); // update timedtamp of last measurement
+		unsigned channel_index = static_cast<unsigned>(ch);
+
+		if (!_already_connected) {
+			_already_connected = true;
+			PX4_INFO("Device ready");
+		}
+
+		if (channel_index < PX4_MAX_ADC_CHANNELS) {
+			_regulator_report.channel_id[channel_index] = static_cast<int>(channel_index);
+			_regulator_report.raw_data[channel_index] = value;
+		}
+
+	} else {
+		// check for timeout
+		if (hrt_elapsed_time(&_last_successful_measurement) > MEASUREMENT_TIMEOUT_US) {
+			// Print message once on disconnect
+			if (_already_connected) {
+				PX4_ERR("No valid measurements for %.1f seconds - device may be disconnected", MEASUREMENT_TIMEOUT_US / 1e6);
+				_already_connected = false;
+			}
+
+			perf_count(_comms_errors);
+		}
+	}
+
+	// Only publish if we have recent data
+	if (hrt_elapsed_time(&_last_successful_measurement) <= MEASUREMENT_TIMEOUT_US) {
+
+		// Publish after all channels have been sampled
+		if (++_channel_cycle_count >= 8) {
+			_channel_cycle_count = 0;
+			_regulator_report_pub.publish(_regulator_report);
+		}
+	}
+
+	perf_end(_cycle_perf);
+}
 
 int ADS7830::init()
 {
@@ -57,7 +135,7 @@ int ADS7830::init()
 	}
 
 	px4_usleep(10000); // Sleep for 10ms after turning REF ON
-	PX4_INFO("succesfully init i2c");
+	PX4_DEBUG("succesfully init i2c");
 
 	ScheduleOnInterval(SAMPLE_INTERVAL / 4, SAMPLE_INTERVAL / 4);
