+++ src/modules/logger/logged_topics.cpp
@@ -117,6 +117,7 @@ void LoggedTopics::add_default_topics()
 	add_optional_topic("rover_velocity_status", 100);
 	add_topic("rtl_time_estimate", 1000);
 	add_topic("rtl_status", 2000);
+	add_optional_topic("regulator_report", 100);
 	add_optional_topic("sensor_airflow", 100);
 	add_topic("sensor_combined");
 	add_optional_topic("sensor_correction");
